+++ src/components/gmdaggerheart/Token/Attack.tsx
@@ -1,7 +1,4 @@
-import { IDiceRoll, Operator, parseRollEquation } from "dddice-js";
-import { getUserUuid, localRoll, rollerCallback, rollWrapper } from "../../../helper/diceHelper.ts";
-import { updateTokenMetadata } from "../../../helper/tokenHelper.ts";
-import { updateRoomMetadata } from "../../../helper/helpers.ts";
+import { rollDualityDice } from "../../../helper/diceHelper.ts";
 import { useDiceRoller } from "../../../context/DDDiceContext.tsx";
 import { useShallow } from "zustand/react/shallow";
 import { useMemo, useState } from "react";
@@ -48,104 +45,12 @@ export const Attack = ({ id, character }: { id: string; character: string }) => 
     const hover = useHover(context, { handleClose: safePolygon() });
 
     const { getReferenceProps, getFloatingProps } = useInteractions([hover]);
-    const roll = async (modifier?: "ADV" | "DIS" | "SELF") => {
-        let parsedDice: Array<{
-            dice: IDiceRoll[];
-            operator: Operator | undefined;
-        }> = [];
-
-        if (!room?.disableDiceRoller) {
-            parsedDice.push(parseRollEquation("1d12", hopeTheme ? hopeTheme.id : theme?.id || "dddice-bees"));
-            parsedDice.push(parseRollEquation("1d12", fearTheme ? fearTheme.id : theme?.id || "dddice-bees"));
-
-            const parsed: {
-                dice: IDiceRoll[];
-                operator: Operator | undefined;
-            } = { dice: [], operator: undefined };
-            parsedDice.forEach((p) => {
-                parsed.dice.push(...p.dice);
-            });
-
-            if (modifier === "ADV") {
-                parsed.dice.push({ type: "d6", theme: theme?.id || "dddice-bees" });
-            } else if (modifier === "DIS") {
-                parsed.dice.push({ type: "d6", theme: theme?.id || "dddice-bees" });
-                parsed.operator = { "*": { "-1": [2] } };
-            }
-
-            if (bonus !== 0) {
-                parsed.dice.push({ type: "mod", value: bonus });
-            }
-
-            if (rollerApi) {
-                try {
-                    let rollResult = await rollWrapper(rollerApi, parsed.dice, {
-                        label: "Attack",
-                        operator: parsed.operator,
-                        external_id: character,
-                        whisper: modifier === "SELF" ? await getUserUuid(room, rollerApi) : undefined,
-                    });
-                    if (rollResult) {
-                        if (rollResult.values[0].value > rollResult.values[1].value) {
-                            rollResult.label += ": Hope";
-                            await updateTokenMetadata({ ...data, hope: Math.min(data.hope + 1, 6) }, [id]);
-                        } else if (rollResult.values[0].value < rollResult.values[1].value) {
-                            rollResult.label += ": Fear";
-                            await updateRoomMetadata(room, { fear: room?.fear ? Math.min(room?.fear + 1, 12) : 1 });
-                        } else {
-                            rollResult.label += ": Critical";
-                            await updateTokenMetadata(
-                                {
-                                    ...data,
-                                    hope: Math.min(data.hope + 1, 6),
-                                    stress: { ...data.stress, current: Math.max(data.stress.current - 1, 0) },
-                                },
-                                [id],
-                            );
-                        }
-                        await rollerCallback(rollResult, addRoll);
-                    }
-                } catch {
-                    console.warn("error in dice roll", parsed.dice, parsed.operator);
-                }
-            }
-        } else {
-            let notation = "2d12";
-            if (modifier === "ADV") {
-                notation += "+1d6";
-            } else if (modifier === "DIS") {
-                notation += "-1d6";
-            }
-            if (bonus !== 0) {
-                notation += `+ ${bonus}`;
-            }
-
-            const result = await localRoll(notation, "Attack", addRoll, modifier === "SELF", character, true);
-
-            // @ts-ignore
-            const rolls: Array<{ value: number }> = result?.rolls[0].rolls;
-            if (rolls && rolls[0].value < rolls[1].value) {
-                await updateRoomMetadata(room, { fear: room?.fear ? Math.min(room?.fear + 1, 12) : 1 });
-            } else if (rolls && rolls[0].value > rolls[1].value) {
-                await updateTokenMetadata({ ...data, hope: Math.min(data.hope + 1, 6) }, [id]);
-            } else {
-                await updateTokenMetadata(
-                    {
-                        ...data,
-                        hope: Math.min(data.hope + 1, 6),
-                        stress: { ...data.stress, current: Math.max(data.stress.current - 1, 0) },
-                    },
-                    [id],
-                );
-            }
-        }
-    };
 
     const getDicePreview = () => {
         try {
             if (theme) {
                 const hope = getDiceImage(hopeTheme ?? theme, "d12", 0);
-                const fear = getDiceImage(fearTheme ?? theme, "d12", 0);
+                const fear = getDiceImage(fearTheme ?? theme, "d12", 1);
                 return (
                     <>
                         {hope}
@@ -155,16 +60,16 @@ export const Attack = ({ id, character }: { id: string; character: string }) => 
             } else {
                 return (
                     <>
-                        <D12 />
-                        <D12 />
+                        <D12 key={"hope"} />
+                        <D12 key={"fear"} />
                     </>
                 );
             }
         } catch {
             return (
                 <>
-                    <D12 />
-                    <D12 />
+                    <D12 key={"hope"} />
+                    <D12 key={"fear"} />
                 </>
             );
         }
@@ -180,7 +85,19 @@ export const Attack = ({ id, character }: { id: string; character: string }) => 
                     className={`dice-button button ${rolling ? "rolling" : ""} ${styles.attackButton}`}
                     onClick={async () => {
                         setRolling(true);
-                        await roll();
+                        await rollDualityDice(
+                            bonus,
+                            "Attack",
+                            character,
+                            data,
+                            id,
+                            addRoll,
+                            room,
+                            hopeTheme,
+                            fearTheme,
+                            theme,
+                            rollerApi,
+                        );
                         setRolling(false);
                     }}
                 >
@@ -211,7 +128,20 @@ export const Attack = ({ id, character }: { id: string; character: string }) => 
                         className={"advantage"}
                         disabled={!isEnabled}
                         onClick={async () => {
-                            await roll("ADV");
+                            await rollDualityDice(
+                                bonus,
+                                "Attack",
+                                character,
+                                data,
+                                id,
+                                addRoll,
+                                room,
+                                hopeTheme,
+                                fearTheme,
+                                theme,
+                                rollerApi,
+                                "ADV",
+                            );
                         }}
                     >
                         ADV
@@ -220,17 +150,64 @@ export const Attack = ({ id, character }: { id: string; character: string }) => 
                         className={"disadvantage"}
                         disabled={!isEnabled}
                         onClick={async () => {
-                            await roll("DIS");
+                            await rollDualityDice(
+                                bonus,
+                                "Attack",
+                                character,
+                                data,
+                                id,
+                                addRoll,
+                                room,
+                                hopeTheme,
+                                fearTheme,
+                                theme,
+                                rollerApi,
+                                "DIS",
+                            );
                         }}
                     >
                         DIS
                     </button>
-
+                    <button
+                        className={"reaction"}
+                        disabled={!isEnabled}
+                        onClick={async () => {
+                            await rollDualityDice(
+                                bonus,
+                                "Attack",
+                                character,
+                                data,
+                                id,
+                                addRoll,
+                                room,
+                                hopeTheme,
+                                fearTheme,
+                                theme,
+                                rollerApi,
+                                "REACT",
+                            );
+                        }}
+                    >
+                        REACT
+                    </button>
                     <button
                         className={"self"}
                         disabled={!isEnabled}
                         onClick={async () => {
-                            await roll("SELF");
+                            await rollDualityDice(
+                                bonus,
+                                "Attack",
+                                character,
+                                data,
+                                id,
+                                addRoll,
+                                room,
+                                hopeTheme,
+                                fearTheme,
+                                theme,
+                                rollerApi,
+                                "SELF",
+                            );
                         }}
                     >
                         HIDE