+++ src/components/gmdaggerheart/Token/Stats.tsx
@@ -2,12 +2,11 @@ import { GMDMetadata } from "../../../helper/types.ts";
 import { useDiceRoller } from "../../../context/DDDiceContext.tsx";
 import { useShallow } from "zustand/react/shallow";
 import { Image } from "@owlbear-rodeo/sdk";
-import { getTokenName, updateRoomMetadata } from "../../../helper/helpers.ts";
+import { getTokenName } from "../../../helper/helpers.ts";
 import { useEffect, useMemo, useState } from "react";
 import { autoPlacement, safePolygon, useFloating, useHover, useInteractions } from "@floating-ui/react";
 import { useMetadataContext } from "../../../context/MetadataContext.ts";
-import { IDiceRoll, Operator, parseRollEquation } from "dddice-js";
-import { getUserUuid, localRoll, rollerCallback, rollWrapper } from "../../../helper/diceHelper.ts";
+import { rollDualityDice } from "../../../helper/diceHelper.ts";
 import { useRollLogContext } from "../../../context/RollLogContext.tsx";
 import styles from "./stats.module.scss";
 import { EditSvg } from "../../svgs/EditSvg.tsx";
@@ -68,107 +67,6 @@ export const Stat = ({
 
     const { getReferenceProps, getFloatingProps } = useInteractions([hover]);
 
-    const roll = async (modifier?: "ADV" | "DIS" | "SELF" | "REACT") => {
-        let parsedDice: Array<{
-            dice: IDiceRoll[];
-            operator: Operator | undefined;
-        }> = [];
-
-        if (!room?.disableDiceRoller) {
-            parsedDice.push(parseRollEquation("1d12", hopeTheme ? hopeTheme.id : theme?.id || "dddice-bees"));
-            parsedDice.push(parseRollEquation("1d12", fearTheme ? fearTheme.id : theme?.id || "dddice-bees"));
-
-            const parsed: {
-                dice: IDiceRoll[];
-                operator: Operator | undefined;
-            } = { dice: [], operator: undefined };
-            parsedDice.forEach((p) => {
-                parsed.dice.push(...p.dice);
-            });
-
-            if (modifier === "ADV") {
-                parsed.dice.push({ type: "d6", theme: theme?.id || "dddice-bees" });
-            } else if (modifier === "DIS") {
-                parsed.dice.push({ type: "d6", theme: theme?.id || "dddice-bees" });
-                parsed.operator = { "*": { "-1": [2] } };
-            }
-
-            if (value !== 0) {
-                parsed.dice.push({ type: "mod", value: value });
-            }
-
-            if (rollerApi) {
-                try {
-                    let rollResult = await rollWrapper(rollerApi, parsed.dice, {
-                        label: name,
-                        operator: parsed.operator,
-                        external_id: character,
-                        whisper: modifier === "SELF" ? await getUserUuid(room, rollerApi) : undefined,
-                    });
-                    if (rollResult) {
-                        if (rollResult.values[0].value > rollResult.values[1].value) {
-                            rollResult.label += ": Hope";
-                            if (modifier !== "REACT") {
-                                await updateTokenMetadata({ ...data, hope: Math.min(data.hope + 1, 6) }, [id]);
-                            }
-                        } else if (rollResult.values[0].value < rollResult.values[1].value) {
-                            rollResult.label += ": Fear";
-                            if (modifier !== "REACT") {
-                                await updateRoomMetadata(room, { fear: room?.fear ? Math.min(room?.fear + 1, 12) : 1 });
-                            }
-                        } else {
-                            rollResult.label += ": Critical";
-                            if (modifier !== "REACT") {
-                                await updateTokenMetadata(
-                                    {
-                                        ...data,
-                                        hope: Math.min(data.hope + 1, 6),
-                                        stress: { ...data.stress, current: Math.max(data.stress.current - 1, 0) },
-                                    },
-                                    [id],
-                                );
-                            }
-                        }
-                        await rollerCallback(rollResult, addRoll);
-                    }
-                } catch {
-                    console.warn("error in dice roll", parsed.dice, parsed.operator);
-                }
-            }
-        } else {
-            let notation = "2d12";
-            if (modifier === "ADV") {
-                notation += "+1d6";
-            } else if (modifier === "DIS") {
-                notation += "-1d6";
-            }
-            if (value !== 0) {
-                notation += `+ ${value}`;
-            }
-
-            const result = await localRoll(notation, name, addRoll, modifier === "SELF", character, true);
-
-            // @ts-ignore
-            const rolls: Array<{ value: number }> = result?.rolls[0].rolls;
-            if (modifier !== "REACT") {
-                if (rolls && rolls[0].value < rolls[1].value) {
-                    await updateRoomMetadata(room, { fear: room?.fear ? Math.min(room?.fear + 1, 12) : 1 });
-                } else if (rolls && rolls[0].value > rolls[1].value) {
-                    await updateTokenMetadata({ ...data, hope: Math.min(data.hope + 1, 6) }, [id]);
-                } else {
-                    await updateTokenMetadata(
-                        {
-                            ...data,
-                            hope: Math.min(data.hope + 1, 6),
-                            stress: { ...data.stress, current: Math.max(data.stress.current - 1, 0) },
-                        },
-                        [id],
-                    );
-                }
-            }
-        }
-    };
-
     return (
         <div className={`button-wrapper ${styles.stat}`}>
             <span className={styles.name}>{name.substring(0, 3)}</span>
@@ -192,7 +90,19 @@ export const Stat = ({
                     className={`dice-button button ${rolling ? "rolling" : ""} ${styles.statValue}`}
                     onClick={async () => {
                         setRolling(true);
-                        await roll();
+                        await rollDualityDice(
+                            value,
+                            name,
+                            character,
+                            data,
+                            id,
+                            addRoll,
+                            room,
+                            hopeTheme,
+                            fearTheme,
+                            theme,
+                            rollerApi,
+                        );
                         setRolling(false);
                     }}
                 >
@@ -210,7 +120,20 @@ export const Stat = ({
                         className={"advantage"}
                         disabled={!isEnabled}
                         onClick={async () => {
-                            await roll("ADV");
+                            await rollDualityDice(
+                                value,
+                                name,
+                                character,
+                                data,
+                                id,
+                                addRoll,
+                                room,
+                                hopeTheme,
+                                fearTheme,
+                                theme,
+                                rollerApi,
+                                "ADV",
+                            );
                         }}
                     >
                         ADV
@@ -219,7 +142,20 @@ export const Stat = ({
                         className={"disadvantage"}
                         disabled={!isEnabled}
                         onClick={async () => {
-                            await roll("DIS");
+                            await rollDualityDice(
+                                value,
+                                name,
+                                character,
+                                data,
+                                id,
+                                addRoll,
+                                room,
+                                hopeTheme,
+                                fearTheme,
+                                theme,
+                                rollerApi,
+                                "DIS",
+                            );
                         }}
                     >
                         DIS
@@ -229,7 +165,20 @@ export const Stat = ({
                             className={"reaction"}
                             disabled={!isEnabled}
                             onClick={async () => {
-                                await roll("REACT");
+                                await rollDualityDice(
+                                    value,
+                                    name,
+                                    character,
+                                    data,
+                                    id,
+                                    addRoll,
+                                    room,
+                                    hopeTheme,
+                                    fearTheme,
+                                    theme,
+                                    rollerApi,
+                                    "REACT",
+                                );
                             }}
                         >
                             REACT
@@ -239,7 +188,20 @@ export const Stat = ({
                         className={"self"}
                         disabled={!isEnabled}
                         onClick={async () => {
-                            await roll("SELF");
+                            await rollDualityDice(
+                                value,
+                                name,
+                                character,
+                                data,
+                                id,
+                                addRoll,
+                                room,
+                                hopeTheme,
+                                fearTheme,
+                                theme,
+                                rollerApi,
+                                "SELF",
+                            );
                         }}
                     >
                         HIDE